switching to using messages_count instead
Because turned out we can't depend on messages_count_unread. Because the only way to make a message as read is by making an API call using that specific message id. So if we get a notification saying that our prayer got prayed for 5 times. Then we need the ids of those 5 messages to dismiss the notification. Even if we say mark the last 5 messages as read, then we will face a problem with concurrency (Another user might have prayed again already)